release: v0.8.17 - topic dedup refactor, v4v bolt11 amount fix, and changelog
- Bump version to 0.8.17 across Cargo.toml, package.json, README.md, Cargo.lock
- Extract duplicate topic post merge logic into topic_store::merge_pending_posts
- Fix v4v payment using parse_bolt11_amount_msats instead of parse_bolt11_amount
  for correct msat comparison in lnurl invoice validation
- Add changelog for 0.8.17 release
