+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nostrblue"
-version = "0.8.16"
+version = "0.8.17"
 edition = "2021"
 rust-version = "1.82"
 authors = ["Patrick Ulrich"]
