+++ public/docs/changelogs/0.8.17.md
@@ -1,0 +1,11 @@
+- Followers/Following Modal
+- Relay-Aware Live Chat
+- NIP-73 Web URL Topics with real-time updates
+- DB-First Note Threads with NIP-22 comment support
+- Gap-Aware Feed Pagination
+- Nests audio engine rewrite
+- Cashu wallet seed zeroization
+- Bolt11 invoice parsing fix
+- Blossom server URL validation
+- Subscription leak and notification lifecycle fixes
+- Bug fixes and performance improvements
